Change navigation functionality Fixed working with the database (added the function of taking a player by ID)

+++ app/src/main/java/com/example/footballplayers/domain/usecases/DeletePlayerUseCase.kt
@@ -3,7 +3,7 @@ package com.example.footballplayers.domain.usecases
 import com.example.footballplayers.domain.repository.PlayerRepository
 
 class DeletePlayerUseCase(private val playerRepository: PlayerRepository) {
-    suspend fun execute(id : Int) {
+    suspend fun execute(id : Long) {
         playerRepository.deletePlayer(id)
     }
 }
