+++ app/src/main/java/com/example/footballplayers/domain/usecases/EditPlayerUseCase.kt
@@ -4,7 +4,7 @@ import com.example.footballplayers.domain.repository.PlayerRepository
 
 class EditPlayerUseCase(private val playerRepository: PlayerRepository) {
 
-    suspend fun execute(id : Int, newFirstName : String, newLastName : String) {
+    suspend fun execute(id : Long, newFirstName : String, newLastName : String) {
         if(newFirstName != "" && newLastName != "") {
             playerRepository.updatePlayer(id, newFirstName, newLastName)
         }
